"Add: OnClick Event that triggers title change on each category display"
A State was created.
A function, titleHandler, was passed down all the way down to the lowest child component, Checkbox.
A variable, title, was passed down to another child component, Order.
When the buttons in Checkbox component receive a click, the title in the Order component changes, keeping it consistent with the category name.